+++ codegen/src/main/kotlin/software/amazon/smithy/rust/codegen/smithy/protocols/serialize/JsonSerializerGenerator.kt
@@ -244,6 +244,7 @@ class JsonSerializerGenerator(protocolConfig: ProtocolConfig) : StructuredDataSe
             is NumberShape -> {
                 val numberType = when (symbolProvider.toSymbol(target).rustType()) {
                     is RustType.Float -> "Float"
+                    // NegInt takes an i64 while PosInt takes u64. We need this to be signed here
                     is RustType.Integer -> "NegInt"
                     else -> throw IllegalStateException("unreachable")
                 }
